// apps/mobile/src/ui/theme/tokens.generated.ts
export const tokens = {
  color: {
    primitive: {
      blue: {
        '50': '#e8f1fe',
        '100': '#d9e6ff',
        '200': '#c5d7fb',
        '300': '#9db7f9',
        '400': '#7096f8',
        '500': '#4979f5',
        '600': '#3460fb',
        '700': '#264af4',
        '800': '#0031d8',
        '900': '#0017c1',
        '1000': '#00118f',
        '1100': '#000071',
        '1200': '#000060',
      },
// ...
    },
// ...
    semantic: {
      success: {
        '1': '#259d63',
        '2': '#197a4b',
      },
      error: {
        '1': '#ec0000',
        '2': '#ce0000',
      },
      warning: {
        yellow: {
          '1': '#b78f00',
          '2': '#927200',
        },
        orange: {
          '1': '#fb5b01',
          '2': '#c74700',
        },
      },
    },
  },
// ...
} as const;